+++ server/app/models/user.go
@@ -1,6 +1,7 @@
 package models
 
 import (
+	"encoding/json"
 	"fmt"
 	"log"
 	"net/http"
@@ -59,4 +60,49 @@ func FindUserByID(r *http.Request) (entity.User, error) {
 		return user, err
 	}
 	return user, nil
+}
+
+func UpdateUser(foundUser entity.User, r *http.Request) (err error) {
+	type body struct {
+		User struct {
+			Password string `json:"password,omitempty"`
+		} `json:"user,omitempty"`
+		NewUser struct {
+			UserName string `json:"user_name,omitempty"`
+			Email    string `json:"email,omitempty"`
+			Password string `json:"password,omitempty"`
+		} `json:"new_user,omitempty"`
+	}
+
+	var updateUser body
+	json.NewDecoder(r.Body).Decode(&updateUser)
+
+	if err := bcrypt.CompareHashAndPassword([]byte(foundUser.Password), []byte(updateUser.User.Password)); err != nil {
+		return err
+	}
+
+	if updateUser.NewUser.UserName != "" {
+		foundUser.UserName = updateUser.NewUser.UserName
+	}
+	if updateUser.NewUser.Email != "" {
+		foundUser.Email = updateUser.NewUser.Email
+	}
+	if updateUser.NewUser.Password != "" {
+		hash, err := bcrypt.GenerateFromPassword([]byte(updateUser.NewUser.Password), 10)
+		if err != nil {
+			log.Fatal(err)
+		}
+		foundUser.Password = string(hash)
+	}
+
+	db, err := db.SQLConnect()
+	if err != nil {
+		panic(err.Error())
+	}
+	defer db.Close()
+	if err := db.Save(&foundUser).Error; err != nil {
+		return err
+	}
+
+	return nil
 }
